Throw exceptions for getListenerPriority() incompatibilities
For the time being, do not support wildcard patterns with getListenerPriority().

An exception will be thrown if getListenerPriority() is not implemented on the inner dispatcher. The method was added in https://github.com/symfony/symfony/pull/16198 for Symfony 2.8. It was added to the interface in Symfony 3.0, which is relevant for our test cases.